neural-style-transfer/requirements.txt
gradio==3.50.2
huggingface-hub==0.16.4
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.24.0
